Create Make-Import-ExchangeCert.ps1
This can run as a scheduled task and import an ACME certificate made by Certbot into Exchange.  It will also delete the old certificate.  That's so you don't have them piling up in your certificate store.   It will send an email message indicating whether or not it has been successful.